ref(useGenerativePreview.ts): move  composable to utils
